2.5 - Separated history component
