+++ src/app/app.router.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { HistoryComponent } from './components/youtube-history.component';
+import { HistoryComponent } from './components/history/history.component';
 import { AboutComponent } from './components/about/about.component';
 import { SettingsComponent } from './components/settings/settings.component';
 import { CategoryComponent } from './components/category/category.component';
