+++ src/app/components/history/history.component.html
@@ -1,0 +1,20 @@
+<div class="app-head">
+    <h2>History</h2>
+</div>
+<div class="app-content">
+    <div class="history-video-content">
+        <div *ngIf="globals.historyVideos.length === 0" class="video-list-info">
+            No history
+        </div>
+        <div *ngFor="let historyVideo of globals.historyVideos; let i = index" [attr.data-index]="i" class="video-item">
+            <div class="video-item-info">
+                <div *ngIf="globals.thumbnails" class="video-item-image" [ngStyle]="{'background-image': 'url(' + historyVideo.thumbnails.default +')'}"></div>
+                <div class="video-item-content">
+                <p class="video-item-title">{{historyVideo.title}}</p>
+                <p class="video-item-autor">{{historyVideo.channelTitle}}</p>                      
+                </div>
+            </div>
+            <app-buttons [videoIndex]="i" [listID]="4"></app-buttons>
+        </div>
+    </div>
+</div>
