+++ src/app/components/history/history.component.scss
@@ -1,0 +1,3 @@
+.history-video-content .video-item {
+    margin-bottom: 10px;
+}
