+++ src/assets/scss/_videos.scss
@@ -31,7 +31,3 @@
         margin-right: 3px;
     }
 }
-
-.history-video-content .video-item {
-    margin-bottom: 10px;
-}
